cr50: add vendor command for retrieving flash elog
We want to be able to retrieve flash log contents using the vendor
command channel.

The input parameter of the command is the timestamp of the last
retrieved log event, or zero of the AP wants to start over.

The response is the next entry after the one requested by the AP, or
an empty message if there are no newer log messages.

BRANCH=cr50, cr50-mp
BUG=b:63760920
TEST=with the upcoming gsctools changes observed the ability to
     retrieve flash log messages.

Change-Id: I7a5438daab780c80f77cc7ebda5b719814b46489
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1525146
Reviewed-by: Namyoon Woo <namyoon@chromium.org>
(cherry picked from commit 705595bb681f362a5e34fb377d21fa8bfa605f9d)