ish: Add watchdog timer
This adds support for the watchdog timer (WDT) available on Intel Sensor
Hub (ISH). The WDT default implementation is to warn after 9.5 seconds,
and kill after 0.5 seconds.

Currently, we are waiting on b:128679825 to find out if we can configure
the WDT kill to reset the Minute-IA core rather than kill the entire
notebook. In addition, we want to know if there is a mechanism to
implement max resets (to prevent a boot loop).

BUG=b:127980538,b:128679825
BRANCH=none
TEST=Used waitms command on arcada to verify WDT triggered a warning
IRQ after T1, and killed the system after T2.

Change-Id: I4bd16c253110d60c57eb24cda2abc0facee20748
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>