Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> > Patch Set 3: Code-Review+2 Commit-Queue+1
> > 
> > > Patch Set 3: -Commit-Queue
> > > 
> > > Commit-Queue label set to 0 programmatically by chromite.
> > 
> > chromeos-ec-0.0.1-r5420: /opt/coreboot-sdk/lib/gcc/nds32le-elf/8.2.0/../../../../nds32le-elf/bin/ld: build/bip/RO/ec.RO.elf section `.data' will not fit in region `FLASH'
> > chromeos-ec-0.0.1-r5420: /opt/coreboot-sdk/lib/gcc/nds32le-elf/8.2.0/../../../../nds32le-elf/bin/ld: region `FLASH' overflowed by 24 bytes
> > 
> > buildall passes. Are we building these differently (using coreboot-sdk) on builders?
> 
> Maybe this change plus other in the CQ run went over the limit. We should start remove command from bip at this point since no one really uses it.

this also happens when it's building just this change, which makes it weird - as if that builder fails to build bip as soon as EC needs to be rebuilt at all. (and why are we running tests for all boards on all builders? seems wasteful to me)

Patch-set: 3
Reviewer: Gerrit User 1118539 <1118539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>