cr50: add support for building CR50_SQA images
Devices in the lab need to be switched from prepvt to pvt flags. This
change adds a build flag for CR50_SQA. These images have more
capabilities than normal cr50 images, but less than CR50_DEV images.

SQA images can access the rollback command, update's will not reject
images for any reason, and 'bid force_pvt' can be used to set the bid
flags to 0. bid force_pvt does not erase flashinfo. After the board id
has been set, we can still change 1 bits to 0, so we don't need to
eraseflashinfo change the board id flags to 0. This makes the command a
bit safer, because the board id RLZ can't be changed just the flags.
Changing the flags to 0 works for the test team, because it prevents
cr50 from updating to prepvt.

This change also changes marks rollback as a safe command. CR50_SQA
images aren't automatically open, so if rollback isn't safe, they will
have to go through the open process to change to pvt. Rollback is only
included in DEV and SQA images, so it's ok if it's marked safe. It's
understood these images aren't supposed to be used on regular devices.
They're just for special development and test environments.

NEW PROCESS FOR SQA:
- update to sqa image
- run 'bid force_pvt'
- flash pvt image
- run rollback

BRANCH=none
BUG=b:126618143
TEST=run sqa process

Change-Id: Ia713274830c9e19cdb3ccafa87849c160d667683
Signed-off-by: Mary Ruthven <mruthven@google.com>