pi3usb9281: Detect detach event by interrupt register
Currently, when device is attached or detached, bc12_detect will reset
the chip and let it run the detection again.

This patch checks the interrupt register first on INTB trigger. If the
detach bit is set, it calls usb_charger_reset_charge. This is based on
the observation that the interrupt register correctly registers the
detach bit right after INTB trigger.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/125176293
BRANCH=Nami
TEST=Verify a wall charger is detected as 3 and Syndra is detected as
5 consistently. Detach event is correctly caught.

Change-Id: Icb6b49a03190d21194526429fc214969d7f46fea