Update patch set 4
Patch Set 4:

(2 comments)

Sorry, meant to answer this, but it slipped through the cracks:

> > 
> > If I understand the bug correctly, how do plan to use this to save panic data? I'm a bit worried that we might *not* be able to run such complex code in the EC panic path.
> > 

good point, this definitely is not going to be usable for panics. Do you want me to mention this in the bug or somewhere in the code?

> > Also, are we concerned about flash maximum erase cycles?
>

not really, this log is supposed to be growing very slowly, so erase cycles will be rare in normal conditions. If flash gets worn down because of this we sure have big problems to solve apart from log flash wear.
 
> I had 2 questions above ^^

Patch-set: 4
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>