Nami: Allow Ekko, Bard, Pyke to defer enabling BC 1.2
Currently, Nami variants prevents AP from booting when the battery is
discharged and the power supply is not more than 15W. A user has to
wait for a battery to be charged, which takes several minutes.

This patch reduces the AP boot threshold to 15W. This is possible
because BC 1.2 controllers can be now enabled or disabled by the EC.

When a battery is discharged, the following sequence is expected:

1. When EC starts up, BC1.2 controllers are disabled.
2. EC boots AP when 15W AC power is available.
3. BIOS verifies EC-RW and EC jumps to RW.
4. EC-RW negotiates higher power from USBC-C charger.
5. BIOS asks EC whether OS power is ready in a loop.
6. EC checks AC power.
  If it's enough, it enables BC1.2.
  If not, returns NACK to BIOS. (Go to step 5)
7. BIOS verifies and boots kernel

Note 1: EC enables BC 1.2 unconditionally in recovery mode.
Note 2: Once BC1.2 is enabled, it won't be disabled until EC resets.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/122896801,b/111990386,b/112172032
BRANCH=nami
TEST=Boot Ekko on discharged battery with USB-C charger.
TEST=Verify USB flash drive works in recovery mode.
TEST=Verify USB keyboard works.
TEST=Verify keyboard backlight works.

Change-Id: Ie69e3dc6cb0a02c26f7d8c07baf5ad631ac282eb