ampton: turn the USB type-a charging off by default in suspend mode
This patch turns the USB type-a charging off by default in suspend mode.
If the DUT power on with connecting a USB device, the OS must send an
event to clear inhibit_charging_in_suspend. Otherwise, the USB device
will lost the power in suspend mode, and the wakeup event won't be
generated if the device is a wakeup source (eg. USB keyboard or mouse).

BUG=b:124698050
BRANCH=octopus
TEST=Power on DUT without connecting USB, then suspend the DUT.
     Verfied the USB type-a charging off.

Change-Id: I9d1b892a40e4d48b93b70d200e3555b04d7dc09b