bip: Delete board
bip has an it8320bx with 256kB of flash space. After dividing by two
and subtracting 0x800, this means the RO section of bip is 126kB.
This is very tight.

We've already removed a few commands to free up space, but this board
hasn't been worked on since summer 2018. Delete it to avoid excessive
maintenance burden.

BUG=b:129283539
BRANCH=none
TEST=make -j buildall

Reviewed-on: https://chromium-review.googlesource.com/1536490
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: Iac903397dd653c8e012c8b3956807ba1bacf681e