Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> verified on bob, reef, and cheza.

the requirement for this to proceed is to audit all released devices schematics to see if the pull up is there, and if not - verify that the EC is pulling up this line, or better yet - test on devices, where external pull up is not present.

Let's keep further discussion on the tracker.

Patch-set: 3
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>