spi: support STM32F412
Enable SPI1_RX and SPI1_TX DMAs when initializing the SPI controller.
The naming conventions are different between the code and the reference
manual.
What is called "channel" in the code is called "stream" in the RM. What
is called "request" in the code is called "channel"(!) in the RM

In the DMA driver, do not overwrite the request in the stream
configuration channel when calling prepare_stream().

BRANCH=None
BUG=b:124996507
TEST=make BOARD=hatch_fp -j
TEST=flash Hatch FPMCU, "ectool --name=cros_fp version" works

Change-Id: I791e39306b088ea28f950d6d5bff11c29828d294
Signed-off-by: Nicolas Norvez <norvez@chromium.org>