bc12: pi3usb9201: Save supplier to update charge manager
When bc1.2 detection is completed, it will update the available charge
for the supplier type which was detected. When bc1.2 is powered down
(upon CC detach), the charge manager needs to be informed that the
available charge for this supplier type is 0,

This CL modifies the pi3usb9201 driver to store the most recent bc1.2
detection supplier type for each port. When the bc1.2 chip is powered
down, then the charge manager can be properly updated.

BUG=b:129435454
BRANCH=none
TEST=Connected, then removed charger to both ports 0 and 1. The
checked status via ectool:

localhost ~ # ectool usbpdpower
Port 0: Disconnected
Port 1: Disconnected

Change-Id: Ic6ae27bb498d2a82389803019f948bc11e3dec23
Signed-off-by: Scott Collyer <scollyer@google.com>