Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Can you remove CQ depend on the kernel repo?

Yup!  I was just mentally noting that on the bus ride in.  There is an upcoming small servod CL to be uploaded and added as CQ-DEPEND here, so I'll update this CL when that one is out for review.

Patch-set: 10
Reviewer: Gerrit User 1291807 <1291807@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>