ish: solve timer reliability and time jumps
b:129805553 demonstrates that the time jumps by 2^32 us twice during
boot. This appears to have been caused by leftover interrupts that
have not been cleared while the system was booting triggering T0.

In addition, b:124890290 discusses that the LRE bit should not be set
in HPET configuration, and the do-while loop introduced in CL:1457597
is no longer needed if the settling spin-lock is placed before arming
the timer. This resolves that

BRANCH=none
BUG=b:124890290,b:12980553
TEST=timer appears to be working reliably for 24 hours on arcada, no
time jumps

Change-Id: Ib0c6b8c9c7e8fe2111a312b9c48061b296178c72
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>