Nami: Enable BC12 when FAFT sets recovery via console
This is a workaround. It should be refined before cherry-picking to ToT.

FAFT emulates recovery boot by 'dut-control power_state:rec', which sets
EC_HOST_EVENT_KEYBOARD_RECOVERY via EC console. This patch adds a check
in the command so that BC 1.2 can be enabled in the emulated recovery
mode as well.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/122896801
BRANCH=nami
TEST=Verify BC1.2 is enabled when 'dut-control power_state:rec' is run.

Change-Id: I120438eea353f66c7a62175d95fcd63bf53e6f45