cr50: add INFO1 RW MAP erasing to CR50_SQA mode
When building with CR50_SQA set to 1 the resulting image is supposed
to allow transitioning the H1 chip it is running on from pre-pvt to MP
track. This should include preserving the Board ID value, but erasing
the board id flags, and erasing the INFO1 RW map. because older MO
image could have a less restrictive mask. The reset of the INFO1 space
should be preserved.

This patch creates a structure describing the layout of the INFO1
space and modifies the 'eraseflashinfo' command to behave differently
depending on the build time configuration. In addition to erasing the
INFO1 RW map:

 - when CR50_DEV is set - everything but INFO1 RO map is erased

 - when CR50_SQA is set - only board ID flags are erased.

With these modifications the 'eraseflashinfo' command is used instead of
'gid force_pvt', and previously erased INFO1 RO map is preserved.

BRANCH=none
BUG=none

TEST=tried running 'eraseflashinfo' in both kinds of images and observed
     desired behavior.

Change-Id: Icf26dc3a4a4bb6fac2fcec630749c81aa46e16ae
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>