USB PD: Send pd_state_names in host command when possible
When the PD state names are available in a build, they are generally
useful to include in the EC_CMD_USB_PD_CONTROL response.  These
strings are picked up in the "ectool usbpd" output, as well as in the
cros_ec_pdinfo which is included in feedback reports.

BUG=None
BRANCH=None
TEST=on casta, ensured that "ectool usbpd" and
/sys/kernel/debug/cros_ec/pdinfo now show state information.  Set
CONFIG_USB_PD_DEBUG_LEVEL to 0 and ensured there was still a reduction
in code size due to the names not being included.

Change-Id: I56aec5c03577d408e90395d4c4877968014100fa
Signed-off-by: Diana Z <dzigterman@chromium.org>