common: Refactor TASK macros to take 5 args instead of 4
The Minute-IA core implementation allows specification of flags as a
fifth argument to the TASK_* macros. This commit refactors all of the
tasklists to:

1. Use 5 arguments instead of four, with a default of 0 for flags
2. Describe the 'f' parameter in the comments
3. Have more consistent formatting between tasklists
4. Correct a documentation typo that existed across most tasklists (in
   -> is)

The corresponding changes to the core/*/task.c files needed were made as
well, and the TEST_TASK_EXTRA_ARGS preprocessor definition was removed
as it is no longer needed.

BUG=chromium:948705
BRANCH=none
TEST=make buildall -j

Change-Id: I70e753a06c2348c4d4772e0004177b02d032d80f
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>