Update patch set 2
Patch Set 2:

Based on discussion with Raul, we may want to abandon this as the extra flag adds quite a bit of noise where it is not needed. Jett: we should talk about a better solution to this.

Patch-set: 2
CC: Gerrit User 1270601 <1270601@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1330845 <1330845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Trybot-Ready=0