cr50: default to SPI PLT_RST if there's a pullup on DIOA9
The PLT_RST_L SYS_RST_L detection is critical to booting, so we need to
have a way to make that decision for SPI boards even if there isn't a
matching config in the board properties table.

We keep adding SPI PLT_RST configs. Everytime we do we have to
immediately add a new config to the strap table, because SPI straps
aren't able to default to PLT_RST_L. This change adds support for
defaulting to PLT_RST_L if there is a pullup on DIOA9. With this new
logic there will be 8 supported configs that default to SPI PLT_RST_L
and 4 that default to SPI SYS_RST_L.

BUG=b:129866334
BRANCH=cr50, cr50-guc
TEST=remove mistral config from the table. Make sure it still defaults
to SPI PLT_RST even though the config isn't found.

Change-Id: If9bc79eccb957d0e3442fd978ef422726b485f0e
Signed-off-by: Mary Ruthven <mruthven@chromium.org>