power: Init host_sleep_event resume_transitions
I had missed a piece of feedback from Karthik about initializing
the sleep transitions member of the resume response to avoid
returning uninitialized data from the EC. This is especially important
if CONFIG_POWER_S0IX_FAILURE_DETECTION is not selected, because then
it will *always* return uninitialized data. This drives the kernel nuts,
since it has no way to know if that config is on or not.

BUG=b:123716513
BRANCH=none
TEST=Test suspend with kernel support for this message, and the
EC config both on and off.

Change-Id: Id35bfe2730327b08f451a4d84277a13e47380a61
Signed-off-by: Evan Green <evgreen@chromium.org>