i2c-pseudo: Simplify install shell script.
The script now unconditionally attempts to (re)install the module, even if
it is already loaded.  (This also removes a message which had a typo,
mistakingly referring to i2c-dev instead of i2c-pseudo.)

The script no longer attempts to load i2c-dev, since servod does that now.

The script now runs "make clean" before "make" .

The header comment no longer references a section of the README that was
never checked in.

BRANCH=none
BUG=79684405
TEST=I ran the install script in various scenarios:
1) No i2c-pseudo module loaded or installed.
2) i2c-pseudo module loaded but not installed.
3) i2c-pseudo module installed but not loaded
4) i2c-pseudo module loaded and installed.

In all cases the module was (re)installed and either left loaded
(if applicable), or loaded (if not previously loaded).

The script does not attempt to unload and re-load the module when
re-installing only for simplicity.

Change-Id: I5f2677b92406b8a49d229bc84d2efcca5504f501
Signed-off-by: Matthew Blecker <matthewb@chromium.org>