ish: reduce hostcmd size to fit within HECI
We only want to send single HECI packets as a response to the AP. 120
bytes should be enough for all commands.

The EC_CMD_CONSOLE_READ Host command was consistently using 2 HECI
packets and this seems to cause issues in the kernel driver.

Also setting a guard variable in a more protected way

BRANCH=none
BUG=b:129937881
TEST=Lock up is much less frequent but not completely gone.

Change-Id: Ifd9dbfdb053fee9894a876fa689f9adf7766fdd5
Signed-off-by: Jett Rink <jettrink@chromium.org>