grunt/usb_pd_policy: Don't disconnect the USB lines while negotiating PD
This policy causes USB 3 to disconnect when doing DP negotiation. If
negotiation results in DOCK then the USB 3 lines are reconnected
resulting in renegotiating the USB device again.

We could make this check the state of the mux and only drop the DP lines
if they are connected. This way keeping the USB lines connected. Though
I'm not sure why we need to disconnect the DP lines in the first place.

BUG=b:123310411
TEST=Tried a test with a monitor and with a USB-C hub with HDMI.
Verified that the extra disconnect is gone.

Change-Id: I907e19689ff608ab2608e973875410cf94bb0053
Signed-off-by: Raul E Rangel <rrangel@chromium.org>