usb_pd+grunt: Wait until PD negotiation has completed to set the MUX
AMD xHCI firmware has a bug where it will dead lock if a device is
disconnected while the SET ADDRESS command is pending. In order to avoid
disconnecting the USB 3 signals we wait to connect them until PD
negotiation has completed.

Worst case this takes 5 seconds because the device does not support PD
so we need to wait until we timeout. Best case it adds about 800ms
between connecting the device and enabling the USB 3 lines.

This patch must be deployed along side kernel changes that block the USB
2 PORT REST until the MUX has been enabled. Otherwise the device will
negotiate as USB 2.

DO NOT SUBMIT: This is probably not the correct way to implement this
CL.

BUG=b:123310411
TEST=Tried with a usb-c monitor, usb-c hub and usb-c to usb 3 adapter.
All devices negotiated correctly.

Change-Id: Iea7817f5c5eabc497abe12e3a995e8237212dd7e
Signed-off-by: Raul E Rangel <rrangel@chromium.org>