mkbp/ish: take timestamps closer to event
For HECI (ISH) take timestamp closer to when the actual hardware
interrupt would be fired on the AP side.

Also only fire the MKBP event on edge transitions.

Detect errors when sending MKBP event and do not honor failed attempts
to flip the state of the MKBP interrupt.

BRANCH=none
BUG=b:129159505
TEST=Fewer MKBP interrupts, timestamps look better, ISH does not get
initially stuck not sending MKBP events

Change-Id: I58522ed52e10f97a465582cd50a11629afa715f9
Signed-off-by: Jett Rink <jettrink@chromium.org>