driver/max695x: Add 7 segment display driver
Port80 information and power sequencing states can be displayed on the
7-segment so that we can observe its information without connecting a
debugger. Adding this CL to initialize and write to the 7-segment display.

For port80, 80(data) gets displayed on the seven segment, for the
power states, EC(state) gets displayed on the seven segment.

The MAX6958/MAX6959 is a compact multiplexed common - cathode display
designed to interface microprocessors with seven - segment numerical LED
digits or to discrete LEDs via a 2-wire serial interface compatible with
SMBusTM and I2C. The 2-wire serial interface uses fixed logic thresholds
of 0.8V/2.1V when the display driver is powered from a 5V supply for
compatibility with 2.5V and 3.3V systems.

BRANCH=None
BUG=b:130738086
TEST=Manually tested on intelrvp, able to write to the 7 segment display

Change-Id: I82ba3aca40640e094463669aa1ff106ce7b99b7b
Signed-off-by: Ayushee <ayushee.shah@intel.com>