flash_ec: Update TODO bug number for servo I2C adapter lookup.
This also removes the placeholder code for a dut-control command that does
not actually exist yet, and this provides a more useful error message when
the current hack for finding the servo I2C adapter fails.

BRANCH=none
BUG=b:79684405,b:130165933
TEST=Ran flash_ec using Servo v4 -> Servo Micro -> Ampton.
With i2c-pseudo loaded and servod running, reflashing worked.
With i2c-pseudo unloaded and servod running, the new error message
was written stderr and flash_ec exited non-zero as desired.

Change-Id: Icfa4f4a894eda6c683a2d1cef4a87b3ac3a0a359
Signed-off-by: Matthew Blecker <matthewb@chromium.org>