util: Add tool to generate cros_ec_commands.h
Add a rule to generate a new cros_ec_commands.h when ec_commands.h is
modified. The rule is checked when buildall is invoked.

At Presubmit stage, check a cros_ec_commands.h exists if ec_commands.h
is modified.

The CL author is responsible to upstream that file.

BUG=chromium:945948
BRANCH=none
CQ-DEPEND=CL:1558853
TEST=Check manually cros_ec_commands.h is generated with
../../repohooks/pre-upload.py
Check checkpatch triggers when it finds an invalid syntax in the output
file.

Change-Id: Ibc8ed7165914d39b5f0bd41643932a8514768925
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>