include: Move RESET_FLAG_* into ec_commands.h as EC_RESET_FLAG_*
RESET_FLAGS_* are used when setting/reading the field ec_reset_flags of
struct ec_response_uptime_info, which is defined in ec_commands.h. So it
might be better to put those macros there.

To be consistent with the other macros in the file, add "EC_" prefixes
to them.

BUG=b:109900671,b:118654976
BRANCH=none
TEST=make buildall -j

Cq-Depend: chrome-internal:1054910,chrome-internal:1054911,
	   chrome-internal:1045539
Change-Id: If72ec25f1b34d8d46b74479fb4cd09252102aafa
Signed-off-by: You-Cheng Syu <youcheng@chromium.org>