Update patch set 2
Abandoned

Raul and I were thinking we should abandon this refactor, as it adds a lot of noise where it is not necessary, and leave it as is (even if kludgy). Something like this could be made significantly better by changing the EC tasklist from preprocessor directives to a YAML (or similar) configuration format, as then parameters could be added as needed without huge refactors like this.

For now, my vote is to abandon.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon