mkbp: handle multiple writes of interrupt
We need to handle the case of multiple tasks trying to set the mkbp
interrupt while the host command task is trying to clear it. The setting
o the interrupt may also take a while and we need to ensure that we
synchronize correct after a longer delay.

BRANCH=none
BUG=b:129159505
TEST=none yet

Change-Id: I056b72c1210d7525c29a8555f97e6f09d773d12f
Signed-off-by: Jett Rink <jettrink@chromium.org>