kukui: Loose charging 5V limitation strategy.
The charging limitation strategy is too conservative: it turns the
charging voltage to 5V once the battery SOC hits 86 percent.  However,
this will make the battery SOC bouncing between 85 and 86 percent
if connected to a external powered hub which can only do 500mA
at 5V (2.5W) and that is not enough to charge the device.

This CL looses the charging limitation strategy by taking IBAT into
consideration.  The strategy is proposed by vendor.  When the battery
is almost fully charged, IBAT will become smaller and smaller.  This CL
sets the IBAT limit to 1A, which is still quite large current, but is
enough to charge battery SOC closing to 99 percent.

It fixes: Battery SOC won't stuck at 85 percent with such hub, now it
can reach 99 percent without damaging the chip.

BRANCH=None
BUG=b:128364884
TEST=- Charge krane with PD charger and see the PD voltage will be
       limited to 5V when SOC > 85% and IBAT <= 1A.
     - When PD voltage been limited to 5V, unplug charger, and wait for
       the SOC drops a little (95 -> 94) then plug a PD charger, and
       see the PD voltage limit been changed to 12.85V.

Change-Id: I7bbdf242ff1568e3595252346eb0a3d7b3ac57ff
Signed-off-by: Yilun Lin <yllin@google.com>