Update patch set 7
Patch Set 7:

> Patch Set 5: Code-Review+2
> 
> I *do* think this is OK as a quick and much needed band-aid.
> 
> Docs don't really say anything about the reliability of HECI, but DCMI-HI (which uses HECI) docs do say this:
> 
>     9.8 Request Retries
>     The DCMI-HI is considered fundamentally reliable. If a correctly formatted request
>     does not receive a response, it is assumed to be for reasons other than data
>     corruption in the transmission of the request or response across the interface.
>     However, the DCMI-HI does include the use of the IPMI Seq field that can be used to
>     differentiate a retried request from an original request. This provides a mechanism
>     that can enable a non-idempotent command to be retried without re-executing it.
>     Refer to the specification of the "BT" interface in [IPMI] for more information on the
>     Seq field and its use.
> 
> So if HECI is "fundamentally reliable", having error handling code like this is inconsequential. However, we should get confirmation from Intel before we keep code like this beyond band-aid basis to make sure there is no use-case for HECI in an unreliable manner.


"no credit" here means that host is not ready to receive. It doesn't matter if the host is not ready due to IRQ delivey problem or Host is really busy with something else.

When host is not ready to receive, we should immediately return to client so give a chance to client if the client wants to retry or revisit with very fresh data.

Blocking for 1 second in this kind common API gives away the chance for a client to do more reasonable action.

This is common function used by clients. 

For HECI bus lockup, currenly, we have http://crrev.com/c/1559848 which it, so far, works fine.

Patch-set: 7
CC: Gerrit User 1137654 <1137654@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>