cr50: casta rbox workaround
Having RBOX INTR_PWRB_IN_FED enabled as a deep sleep wake source on
casta cr50 to hang in brom on deep sleep resume. The power button is not
needed as a wake source on cr50, so this change disables the rbox power
button interrupt when entering deep sleep.

BUG=b:128735316
BRANCH=cr50
TEST=disconnect servo and power. press 'alt + h + vol up' on casta.
Wait 30 seconds for cr50 to enter deep sleep. Use the power button to
wake up the device and make sure it boots.

Change-Id: I02d14e11d9e778e3f44041cf56cc6b6f7a1bef59
Signed-off-by: Mary Ruthven <mruthven@chromium.org>