Update patch set 1
Patch Set 1: Code-Review+1

The changes seems reasonable, and the only 2 places where I see we use it are connect_to_ccd_i2c_bridge() in iteflash and gsctool. And iiuc we want to fail indeed if we found VID/PID, but then failed to connect to a specific i/face & protocol.

The only question: is there a BUG for the issue we are addressing here?

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1