sensor: disable deep sleep while fifo is on
If the sensor fifo is currently on, we expect the EC to be able to
timestamp incoming HW interrupts very quickly. The resume latency for
deep sleep would increase jitter on the EC's timestamp. The fifo should
only be on with the AP is in S0 and an ARC++ app is actively collecting
sensor data.

BRANCH=none
BUG=b:130143396
TEST=ensure that ISH does not go into deep sleep when ARC++ app is open
and using sensors.

Change-Id: I8da9e1c0c28a948a614cd7a32ef53ac23bc76f02
Signed-off-by: Jett Rink <jettrink@chromium.org>