Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why does reset polarity need to be stored in the TCPC config? Isn't that entirely board specific code?

Yes, it is board specific, but so are the other items in the tcpc_config. The first motivation was to be able to just have one implementation for pd_reset_mcu() for all hatch systems, rather than have to keep duplicating this function which will mostly be the same except on Kohaku. Longer term, this could be made into a common function so it doesn't need to be duplicated for each new platform/board.

Patch-set: 2
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>