signer: allow board id of 0:0:0
In the prod branch the prod signer sets the board id to FFFF:0:10000,
because we're going to start doing automatic signing. We need to be able
to override that board id to 0:0:0. bs doesn't support this right now
because it always assumes the first field of the given board id is a 4
character string that it needs to convert into hex. This change modifies
bs to use 0 for the board_id rlz if the rlz is 0 instead of trying to
convert an ascii character into hex. 0 is the only digit bs will accept
for the board_id, because it's easy to interpret.

BUG=none
BRANCH=cr50
TEST=sign image with CR50_BOARD_ID=0:0:0. Make sure gsctool -b shows the
RLZ is 0 instead of 0x30.

Change-Id: I1fe30877ff8d6d1574934c55c2fb5cdc543c34ad
Signed-off-by: Mary Ruthven <mruthven@chromium.org>