atlas: control display backlight from chipset state
this changes how we contorl the display backlight.  previously, we
would enable/disable the backlight master enable directly based on the
lid switch.  with this change, we only change the backlight master
enable when we're notified of the AP entering/leaving suspend.

this helps avoid inconsistencies between the state of the AP and the
state of the backlight.  in particular, if the AP fails to suspend, we
keep the display on so there is visual feedback about the situation.

Change-Id: Ic5c2ab7b4d675f93372d5ced741be90389a9d7f3
Signed-off-by: Caveh Jalali <caveh@chromium.org>