atlas: let AP fully control display backlight
this change removes the EC from the display backlight control logic.
previously, the EC would enable/disable the backlight master enable
directly based on the lid switch.  this turns out to be redundant and
in some cases causes timing violations between backlight enable and
the backlight PWM signal.

the AP already controls the display power load switch
(EN_PP3300_DX_EDP), so when the AP is off, display power (including
the backlight) is also off.  in addition, the AP correctly controls
the backlight using PCH_EDP_BKLTEN.  with this change, we're only
removing the EC's master kill switch for the backlight.

BRANCH=none
BUG=b:129651119
TEST=TBD

Change-Id: Ibef9062ce66dfec1363626e674f144a1a3d55b5e
Signed-off-by: Caveh Jalali <caveh@chromium.org>