Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4: -Code-Review
> > 
> > Please refer to the b:130142249 where notes that there would SKUs without keyboard backlight so please handle this case as well. Thanks.
> 
> Do you mean we need configure PWM off with non-keyboard backlit SKU even HW unstuff the components?

I believe it should be okay to always have the PWM on with the HW unstuffed.  The pwm_channels structure is treated as const, and I don't know that it would be worth losing the const protection on it to let this vary in runtime.

Patch-set: 4
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>