cleanup: update TASK_EVENT_CUSTOM usage
The lightbar.c uses the TASK_EVENT_CUSTOM macro in a non-standard
way comparing it to the rest of the code base. Update the style in
preparation for changing the TASK_EVENT_CUSTOM macro to be a
BUILD_CHECK_INLINE macro. Without this cleanup CL, this code breaks.

BRANCH=none
BUG=none
TEST=builds

Change-Id: I4fdbb35a2aeeb1f8718b22c017607aee5fa1730a
Signed-off-by: Jett Rink <jettrink@chromium.org>