Update patch set 4
Patch Set 4:

i think this patch broke port LEDs on atlas and very likely
may similar devices.
the port no longer lights up when a charger is plugged in.
reverting this patch fixes the issue.

i'm tempted to revert this patch, but i'm not sure that the most
efficient way forward.

Patch-set: 4
CC: Gerrit User 1119124 <1119124@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>