Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Please address comments in the CL so we can submit this change

Here is the register description. I will update register spec later (with other missed information)

The AGENT_STATUS register allows observation of the status of the agent and
attached core. All fields in this register are read-only or write to clear.

29 INBAND_ERROR_SECONDARY Error status for in-band errors with MErrSteer indicating a secondary error
28 INBAND_ERROR_PRIMARY Error status for in-band errors with MErrSteer indicating a primary error (or if error steering is not enabled)
24 MERROR MError assertion detected

For D0ix flow perspective, those status pending bits should be clear to avoid interrupt which blocks the PM flow. So, this CL is needed anyway. For error itself, need to be addressed and see how severe it is. We can do it after merge.

Patch-set: 2
Reviewer: Gerrit User 1314347 <1314347@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>